chore: split object_cache tests into a sibling module (file-size gate)
The #6439 FFI-provenance change grew object_cache.rs to 2035 lines, one
past the 2000-line CI cap (scripts/check_file_size.sh). Move the
`#[cfg(test)] mod object_cache_tests` block (930 lines) verbatim into a
sibling `object_cache/object_cache_tests.rs`, declared `#[cfg(test)] mod
object_cache_tests;`. `super::*` still resolves to the `object_cache`
module, so every test reaches the same private items unchanged. Trunk is
now 1109 lines; no production code moved.
